services/backend/src/pyproject.toml
[project]
name = "fastapi-image-upload-server"
version = "2.0.0"
description = "FastAPI-based image hosting server"
authors = [
    {name = "gxneshx",email = "kalchenko.michael@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "Pillow (>=10.0.0,<11.0.0)",
    "apscheduler (>=3.10.0,<4.0.0)",
    "jinja2 (>=3.1.3,<4.0.0)",
    "slowapi (>=0.1.9,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
